requirements.txt
streamlit
streamlit-antd-components
python-dotenv
supabase